renaming gitea-$ver-linux-$arch to gitea
In gitea.service, the exec is (correctly) /home/git/gitea/gitea web but the downloaded binary was not renamed to gitea and the ervice could not start.  Running the script - as of today - would only lead to gitea-1.0.1-linux-amd64 in the /home/git/gitea/ folder. 

I hope I am right and that this helps.